Analyser.Fun.FunContext.Def

Description

This module defines the function analysis context used during function body type checking.

Function Context

The FunContext extends the global Context with function-specific information:

Local variable scope stack for nested scoping

Current scope level tracking

Expected return type for validation

Scope Management

Uses a NonEmpty stack of scopes where:

Each scope contains local variable declarations

Stack growsshrinks with block entryexit

Variable lookup searches from top to bottom of stack

Lens Integration

Template Haskell generates lenses for convenient field access:

-- Access function context components  
funCtx ^. getCtx         -- Get global context
funCtx ^. getStack       -- Get scope stack
funCtx ^. getLevel       -- Get current scope level
funCtx ^. getRtrnType    -- Get expected return type

-- Update function context
funCtx & getLevel %~ (+1)                    -- Increment scope level
funCtx & getStack %~ (newScope <|)           -- Push new scope

Example Usage

-- Create function context for analysis
let funCtx = FunCtx globalCtx parameterScope 0 TInt

-- Enter new block scope
let newCtx = funCtx & getLevel %~ (+1) 
                    & getStack %~ ([] <|)

-- Exit block scope  
let exitCtx = newCtx & getLevel %~ subtract 1
                     & getStack %~ NE.tail

Documentation

type ScopeLevel = Int

Scope nesting level (0 = function parameter scope, 1+ = nested blocks)

type Scope = [Decl]

A scope containing local variable declarations

data FunContext

Function analysis context with local scope management.

Extends the global Context with function-specific state:

_getCtx

Global analysis context (functions, types, globals, constants)

_getStack

Stack of local scopes (inner scopes at the head)

_getLevel

Current nesting level (0 = function parameters)

_getRtrnType

Expected return type for function validation

The scope stack uses NonEmpty to ensure at least the parameter scope exists. Variable lookup searches from the top (most nested) to bottom (parameters). Just need to verify the type of the vars-4-now Use Lens!!!!!!!! data Context = Ctx { getGlobals :: [GlobalDef] -- could be just the bottom of the scopes , getConsts :: [ConstDef] , getFunDefs :: [FunDef] , getTypeDefs :: [TypeDef] } deriving (Eq, Show)

Constructors

FunCtx

Fields

_getCtx :: Context

Global analysis context (✅ Underscore para lens)

_getStack :: NonEmpty Scope

Local scope stack (innermost first)

_getLevel :: ScopeLevel

Current scope nesting level

_getRtrnType :: Type

Expected function return type

getCtx :: Lens' FunContext Context

getLevel :: Lens' FunContext ScopeLevel

getRtrnType :: Lens' FunContext Type

getStack :: Lens' FunContext (NonEmpty Scope)

Generate lenses for FunContext fields. Creates: getCtx, getStack, getLevel, getRtrnType lens (https:/hackage.haskell.orgpackagelens-tutorial-1.0.5docs/Control-Lens-Tutorial.html):

module Analyser.Context.Def