Grammar.Type

Description

This module defines the complete type system for Moo, including:

Basic primitive types (Int, Char, Float)

Type variables for polymorphism

Parametric data types with constructors

Built-in type patterns with convenient syntax

Type definitions for user-defined types

Example Usage

-- Basic types
intType = TInt
charType = TChar
floatType = TFloat

-- Parametric types
listOfInts = TList TInt
optionalBool = TOption TBool
stringIntPair = TTuple TChar TInt

-- Custom type definition
treeDef = TypeDef Tree ["a"] 
  [ ConstrDef Leaf [TVar "a"]
  , ConstrDef Node [TData Tree [TVar "a"], TVar "a", TData Tree [TVar "a"]]
  ]

Documentation

data Type

Core type system for the Moo language.

Types can be primitive, parametric, or type variables for polymorphism. Type variables are primarily used in type definitions and should be resolved during type checking.

Examples

 -- Primitive types
 intType = TInt
 charType = TChar  
 floatType = TFloat
 
 -- Type variable (for polymorphism)
 genericType = TVar "a"
 
 -- Parametric types
 listType = TData List [TInt]
 treeType = TData Tree [TVar "a", TVar "b"]
 

Constructors

TInt
Integer type

TChar
Character type

TFloat
Floating-point type

TVar Name
Type variable -- .. Only in typedefs, please?

TData Name [Type]
Parametric data type with arguments

data TypeDef

User-defined type with constructors and type parameters.

Type definitions allow creating new algebraic data types with multiple constructors and polymorphic type parameters.

Examples

 -- Simple enumeration
 boolDef = TypeDef Bool [] 
   [ ConstrDef True []
   , ConstrDef False []
   ]
   
 -- Polymorphic option type  
 optionDef = TypeDef Option ["a"]
   [ ConstrDef None []
   , ConstrDef Some [TVar "a"]
   ]
 

Constructors

TypeDef

Fields

tName :: Name

Type name

tParams :: [Name]

Type parameters

tConstrs :: [ConstrDef]

Data constructors

data ConstrDef

Data constructor definition with parameters.

Constructors define how values of a type can be created, specifying the types of arguments they accept.

Examples

 -- Constructor with no arguments
 nilConstr = ConstrDef Nil []
 
 -- Constructor with arguments  
 consConstr = ConstrDef Cons [TVar "a", TList (TVar "a")]
 pairConstr = ConstrDef Pair [TVar "a", TVar "b"]
 

Constructors

ConstrDef

Fields

cName :: Name

Constructor name

cParams :: [Type]

Constructor parameter types

pattern TEmpty :: Type

Built-in type patterns for common types

Empty type (no values)

pattern TOne :: Type

Built-in type patterns for common types

Unit type (single value)

pattern TBool :: Type

Built-in type patterns for common types

Boolean type

pattern TTuple :: Type -> Type -> Type

Tuple type constructor pattern

Tuple type constructor (product type)

pattern TOption :: Type -> Type

Container type patterns

Optional type constructor (sum type)

pattern TList :: Type -> Type

Container type patterns

List type constructor (recursive type)

pattern TDefEmpty :: TypeDef

Pattern synonyms for predefined type definitions

Empty type definition (no constructors, uninhabited type)

pattern TDefOne :: TypeDef

Pattern synonyms for predefined type definitions

Unit type definition (single constructor, single value)

pattern TDefBool :: TypeDef

Pattern synonyms for predefined type definitions

Boolean type definition (two constructors for True/False)

pattern TDefTuple :: TypeDef

Pattern synonyms for predefined type definitions

Tuple type definition (product of two types)

pattern TDefOption :: TypeDef

Pattern synonyms for predefined type definitions

Optional type definition (None or Some value)

pattern TDefList :: TypeDef

Pattern synonyms for predefined type definitions

List type definition (empty list or cons cell)