Moo.GeneticAlgorithm.Random

Description

Some extra facilities to work with Rand monad and PureMT random number generator.

Random numbers from given range

getRandomR :: Random a => (a, a) -> Rand a

Yield a new randomly selected value of type a in the range (lo, hi). See randomR for details.

getRandom :: Random a => Rand a

Yield a new randomly selected value of type a. See random for details.

Probability distributions

getNormal2 :: Rand (Double, Double)

Yield two randomly selected values which follow standard normal distribution.

getNormal :: Rand Double

Yield one randomly selected value from standard normal distribution.

Random samples and shuffles

randomSample :: Int -> [a] -> Rand [a]

Take at most n random elements from the list. Preserve order.

randomSampleIndices :: Int -> Int -> Rand [Int]

Select sampleSize numbers in the range from 0 to (populationSize-1). The function works best when sampleSize is much smaller than populationSize.

shuffle :: [a] -> Rand [a]

Randomly reorder the list.

Building blocks

withProbability :: Double -> (a -> Rand a) -> a -> Rand a

Modify value with probability p. Return the unchanged value with probability 1-p.

Re-exports from random number generator packages

getBool :: Rand Bool

getInt :: Rand Int

getWord :: Rand Word

getInt64 :: Rand Int64

getWord64 :: Rand Word64

getDouble :: Rand Double

runRand

Arguments

:: Rand g a
generator-passing computation to execute

-> g
initial generator

-> (a, g)
return value and final generator

Unwrap a random monad computation as a function. (The inverse of liftRand.)

evalRand

Arguments

:: Rand g a
generator-passing computation to execute

-> g
initial generator

-> a
return value of the random computation

Evaluate a random computation with the given initial generator and return the final value, discarding the final generator.

evalRand m s = fst (runRand m s)

newPureMT :: IO PureMT

Create a new PureMT generator, using the clocktime as the base for the seed.

liftRand

Arguments

:: (g -> (a, g))
pure random transformer

-> Rand g a
equivalent generator-passing computation

Construct a random monad computation from a function. (The inverse of runRand.)

type Rand = Rand PureMT

class Random a

The class of types for which random values can be generated. Most instances of Random will produce values that are uniformly distributed on the full range, but for those types without a well-defined "full range" some sensible default subrange will be selected.

Random exists primarily for backwards compatibility with version 1.1 of this library. In new code, use the better specified Uniform and UniformRange instead.

Since: random-1.0.0

data PureMT

PureMT, a pure mersenne twister pseudo-random number generator